first pr fixes
