+++ .gitignore
@@ -64,3 +64,5 @@ venv.bak/
 !*.ini.dist
 !settings.ini
 !test.ini
+
+eslint.json
