+++ apollo/lib/websocket/app.py
@@ -1,30 +1,11 @@
-import enum
-import json
-
 from fastapi import WebSocket
 from starlette.websockets import WebSocketDisconnect
 
-from apollo.lib.decorators import with_db_session
+from apollo.lib.websocket import ConnectionManager
+from apollo.lib.websocket.interest_type import WebSocketObserverInterestTypes
 
 
-class WebSocketObserverInterestTypes(enum.Enum):
-    @staticmethod
-    @with_db_session
-    def _list_all_agents(session):
-        from apollo.models.agent import list_all_agents
-        from apollo.lib.schemas.agent import BaseAgentSchema
-
-        return [json.loads(BaseAgentSchema.from_orm(agent).json()) for
-                agent in list_all_agents(session)]
-
-    AGENT_LISTING = _list_all_agents
-
-
-class AppConnectionManager:
-    def __init__(self):
-        from apollo.lib.websocket import WebSocketManager
-        self.websocket_manager = WebSocketManager()
-
+class AppConnectionManager(ConnectionManager):
     @staticmethod
     async def _listen(connection: WebSocket):
         try:
@@ -40,7 +21,6 @@ class AppConnectionManager:
         connection_id = await self.websocket_manager.connect_app(
             websocket, connection_type)
         await websocket.send_json(connection_type())
-        await self._listen(websocket)
         await self.websocket_manager.close_app_connection(
             connection_type, connection_id)
